+++ activemodel/lib/active_model/type.rb
@@ -30,9 +30,10 @@ module ActiveModel
         registry.register(type_name, klass, &block)
       end
 
-      def lookup(*args, **kwargs) # :nodoc:
-        registry.lookup(*args, **kwargs)
+      def lookup(*args) # :nodoc:
+        registry.lookup(*args)
       end
+      ruby2_keywords(:lookup)
 
       def default_value # :nodoc:
         @default_value ||= Value.new